Add files via upload
This file contains data on the rodent Hylaeamys megacephalus in a remnant of Brazilian Cerrado. 

Contains data on capture points, sex, weight, morphometric measurements and presence of parasites (warblers, ticks and mites) (2019-2023)

This study was carried out at the Mammal Laboratory of the Federal University of Uberlândia, under the coordination of Natália Leiner (Ethics Committee on Animal Use (CEUA) of the Federal University of Uberlândia (# 041/19).